pyproject.toml
[tool.poetry]
name = "jjaro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.3.0"
numpy = "<2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
